[0x14 ~ 1-print_binary.c] (fix: limit add 40)

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
 * @index: index
 * Return: Always 0 (Success)
 */
-void binary(bool *start_binary, unsigned long int number, int index)
+void binary(bool *start_binary, unsigned long int number, unsigned long int index)
 {
 	if (index == MAX_BINARY)
 		return;
